Refactor hook history cell rendering
